requirements-producer.txt
kafka-python
websockets
asyncio-mqtt
pyyaml
requests